deprecate Base.download in favor of Downloads.download
bump LibCURL to version that doesn't use Base.download
