+++ stdlib/LibCURL.version
@@ -1,2 +1,2 @@
 LIBCURL_BRANCH = master
-LIBCURL_SHA1 = 70aa2cc3f8fa4488897136a0a700b5425bfffdfa
+LIBCURL_SHA1 = 685a4697178415726f1f73d5ccedbc4ed1b5c2ac
